+++ src/Cwaff-Misc/CwaffEnums.cs
@@ -1,0 +1,6 @@
+namespace CwaffingTheGungy;
+
+public static class CwaffGunClass
+{
+    public static readonly GunClass UTILITY = "UTILITY".ExtendEnum<GunClass>();
+}